move all componets in Parent home folder
