+++ src/pages/home/NewsSection.jsx
